+++ scripts/world/terrain_map.gd
@@ -16,12 +16,31 @@ const TERRAIN_ATLAS_COLUMNS := {
 
 @export var automatic_variations := true
 @export var variation_seed := 1979
+@export_range(0.02, 1.0, 0.01) var automatic_refresh_interval := 0.1
+
+var _refresh_queued := false
+var _is_applying_variations := false
+var _refresh_elapsed := 0.0
+var _last_tile_data_hash := 0
+var visual_revision := 0
 
 
 func _ready() -> void:
-	if not Engine.is_editor_hint() and automatic_variations:
-		apply_deterministic_variations()
-	refresh_blending()
+	_last_tile_data_hash = hash(tile_map_data)
+	if not changed.is_connected(_on_terrain_changed):
+		changed.connect(_on_terrain_changed)
+	_schedule_terrain_refresh()
+
+
+func _process(delta: float) -> void:
+	_refresh_elapsed += delta
+	if _refresh_elapsed < automatic_refresh_interval:
+		return
+	_refresh_elapsed = 0.0
+	var current_hash := hash(tile_map_data)
+	if current_hash != _last_tile_data_hash:
+		_last_tile_data_hash = current_hash
+		_schedule_terrain_refresh()
 
 
 func get_terrain_id(cell: Vector2i) -> StringName:
@@ -46,15 +65,19 @@ func paint_terrain(cell: Vector2i, terrain_id: StringName, variation := -1) -> b
 		push_warning("Unknown terrain ID: %s" % terrain_id)
 		return false
 	var selected_variation := variation
-	if selected_variation < 0:
+	if automatic_variations or selected_variation < 0:
 		selected_variation = get_deterministic_variation(cell)
 	selected_variation = posmod(selected_variation, VARIATION_COUNT)
 	set_cell(cell, 0, Vector2i(TERRAIN_ATLAS_COLUMNS[terrain_id], selected_variation), 0)
-	refresh_blending()
+	_last_tile_data_hash = hash(tile_map_data)
+	_schedule_terrain_refresh()
 	return true
 
 
 func apply_deterministic_variations() -> void:
+	if _is_applying_variations:
+		return
+	_is_applying_variations = true
 	for cell in get_used_cells():
 		var atlas_coordinates := get_cell_atlas_coords(cell)
 		if atlas_coordinates.x < 0:
@@ -62,7 +85,8 @@ func apply_deterministic_variations() -> void:
 		var selected_variation := get_deterministic_variation(cell)
 		if atlas_coordinates.y != selected_variation:
 			set_cell(cell, get_cell_source_id(cell), Vector2i(atlas_coordinates.x, selected_variation), get_cell_alternative_tile(cell))
-	refresh_blending()
+	_is_applying_variations = false
+	_last_tile_data_hash = hash(tile_map_data)
 
 
 func get_deterministic_variation(cell: Vector2i) -> int:
@@ -74,3 +98,23 @@ func refresh_blending() -> void:
 	var overlay := get_node_or_null("TerrainBlendOverlay")
 	if overlay != null:
 		overlay.queue_redraw()
+
+
+func _on_terrain_changed() -> void:
+	if not _is_applying_variations:
+		_schedule_terrain_refresh()
+
+
+func _schedule_terrain_refresh() -> void:
+	if _refresh_queued:
+		return
+	_refresh_queued = true
+	call_deferred("_refresh_terrain")
+
+
+func _refresh_terrain() -> void:
+	_refresh_queued = false
+	if automatic_variations:
+		apply_deterministic_variations()
+	refresh_blending()
+	visual_revision += 1